+++ CHANGELOG.md
@@ -11,8 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Translation keys `jarboe::fields.checkbox.yes` and `jarboe::fields.checkbox.no` that are used in `Checkbox` field inline mode.
 - `FieldsetMarkup` field for dividing form into blocks with its own `<legend>`.
 
+### Changed
+- Filter classes method `render` returns `Illuminate\View\View` instance instead of rendered html string.
+
 ### Fixed
 - Immutable `col` value for `Hidden` field.
+- `CheckboxFilter` handle `0` value properly.
 
 ## [1.1.0] - 2020-06-16
 ### Added